requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==8.3.1
numpy==1.21.2
pycryptodome==3.10.1
